Create the CSVWriter class, which builds a record set and then (when flushed) writes the records as a CSV file. Can be used as an *SV class too, you can specify custom delimiters and carriage returns (ie for TSV, tab-separated etc.).
Create the CSVTest class, which instantiates a CSVWriter and then writes some bogus records.

Update the Exam object to have functions to return the raw and percentage score values. TODO: Fix the percentage function to not have the /100 magic number.

Update GradeExams function to put the exam scores into the CSV when grading (in addition to the default printing to console).