GradedExamCollection: update GradedExamCollection to match Grader
GradedExamCollection no longer grades its own exams.

The Grader calls grade() on the exam. This is a good (or less bad) thing (even if it breaks
encapsulation a little) because we want to be able to notify upon a finished grading.

An alternative to consider might be to have GradedExamCollection initialized with an
answerKey and then have an addExam() that takes in an ungraded Exam, scores it, and
adds it to the list, so the Grader loop will look something like:

for page in inputPages:
    exam = examReader.read(exam)
    gradingCollection.addExam(exam)

But for