Add support to exam_report for partial page images!
Since we offer the option to show a partial page, let's allow it. Now just the top 300 pixels of the page will show by default.
This adds methods to InputPage to get the top part as a bufferedImage, and a method to Exam to convert it to a Base64 string,
and then of course the exam report is updated.

The AggregateReport is also tweaked.